Chancepoints

See merge request derdogan18/2021_302_noncompers!26

+++ NeedForSpear/src/domain/obstacle/Obstacle.java
@@ -36,8 +36,8 @@ public abstract class Obstacle extends DomainObject {
 		this.health -= 1;
 		if (this.health == 0) {
 			this.destroy();
-			double o = Game.getInstance().getOldScore();
-			double neww = Game.getInstance().getScore(o);
+			int o = Game.getInstance().getOldScore();
+			int neww = Game.getInstance().getScore(o);
 			Game.getInstance().setScore(neww);
 			System.out.println(Game.getInstance().getOldScore());
 			return true;
